% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitrem_nr.R
\name{fitrem}
\alias{fitrem}
\title{Fit a zero-truncated recurrent events model}
\usage{
fitrem(data, bstart, ...)
}
\arguments{
\item{data}{An array with at least three dimensions}

\item{bstart}{A vector of starting values}
}
\value{
A coefficient matrix, point and interval estimate of the population
  size, loglikelihood and number of iterations
}
\description{
\code{fitrem} fits a zero-truncated recurrent events model on the data
provided and calculates the point and interval estimate of the population
size
}
\details{
This is a specialised function, as it requires the data to be in a
three-dimensional format (persons \times time \times variables). The default
optimization procedure is a Newton-Raphson method, while numerical
optimization via \code{optim} is also possible
}

