% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitrem_nr.R
\name{print.trem}
\alias{print.trem}
\title{Print results of \code{fitrem}}
\usage{
\method{print}{trem}(x, ...)
}
\arguments{
\item{x}{object of class \code{trem}}

\item{...}{Further arguments to be passed from or to other methods}
}
\description{
\code{print.trem} is a S3 method to print the results generated by \code{fitrem}
}

