% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simrem.R
\name{simrem}
\alias{simrem}
\title{Simulate zero-truncated recurrent events data}
\usage{
simrem(N = 5000, time = 365, b0 = -7, bpers = NULL, bage = NULL,
  cyc = F, alpha = NULL, theta = NULL, period = time, int = F,
  aint = NULL, tint = NULL, ...)
}
\arguments{
\item{N}{the total (non-truncated population size)}

\item{time}{the observation window}

\item{b0}{the intercept, default value is -7}

\item{bpers}{coefficients for time-invariant covariates. The data is simulated as ~ B(0.5)}

\item{bage}{coefficients for time-varying dummy variables.}

\item{cyc}{logical, whether cyclical predictors are added to the data. Default is \code{FALSE}}

\item{alpha}{the amplitude of the cyclical effect. Only used when \code{cyc} is \code{TRUE}}

\item{theta}{the phase of the cyclical effect. Only used when \code{cyc} is \code{TRUE}}

\item{period}{the period of the cyclical effect, defaults to \code{time}}

\item{int}{specify whether an interaction between the cyclical effect and the first variable of \code{bpers} should be included}

\item{aint}{the amplitude of the cyclical interaction effect}

\item{tint}{the phase of the cyclical interaction effect}
}
\value{
a 3D-array of dimensions n \times time \times variables. Note that n will vary over each simulation due to sampling fluctuation by drawing from
the total population N
}
\description{
\code{simrem} simulates zero-truncated events data and returns a 3D-array
}

