% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitrem_nr.R
\name{summary.trem}
\alias{summary.trem}
\title{Print a summary of the results from \code{fitrem}}
\usage{
\method{summary}{trem}(x, ...)
}
\description{
\code{summary.trem} is a S3 method that prints more information than \code{print.trem}
}

